various improvements progress bar added

+++ 201-minimal_gnome_desktop.sh
@@ -24,12 +24,33 @@
 # Packages to install
 PACKAGES="gnome-shell gnome-console gnome-tweaks nautilus gnome-disk-utility"
 
+# Function to check if a package is installed
+package_installed() {
+  local package="$1"
+  dpkg -l | grep -q "^ii.*$package"
+}
+
 # Update packages list
 sudo apt-get update &> /dev/null
 
-echo "Install $PACKAGES"
-sudo apt-get install $PACKAGES -y &> /dev/null
+# Check if the packages are already installed
+missing_packages=()
+for pkg in $PACKAGES; do
+  if ! package_installed "$pkg"; then
+    missing_packages+=("$pkg")
+  fi
+done
 
-# Garbage packages cleanup
-# Don't need 'gnome-shell-extension-prefs' as 'gnome-shell-extension-manager' will be installed and performs the same stuff
-sudo apt-get autoremove --purge gnome-shell-extension-prefs -y &> /dev/null
+# Install missing packages
+if [ ${#missing_packages[@]} -gt 0 ]; then
+  echo "Install ${missing_packages[*]}"
+  sudo apt-get install "${missing_packages[@]}" -y &> /dev/null
+
+  # Garbage packages cleanup
+  # Don't need 'gnome-shell-extension-prefs' as 'gnome-shell-extension-manager' will be installed and performs the same stuff
+  sudo apt-get autoremove --purge gnome-shell-extension-prefs -y &> /dev/null
+
+  echo "Installation complete."
+else
+  echo "Packages are already installed."
+fi
